created a few tables in the .sql file
Created the food_order, customer, rewards, and delivery in the DB&tableCreation.sql file. There's an error in the last table so it doesn't run but just to get an update. Once the bug is fixed we can run some queries on it. Here's the link to the tutorial I used:
https://www.youtube.com/watch?v=dMVKy-yTwaM
https://www.youtube.com/watch?v=UQK9_gKQHZg&list=PLyuRouwmQCjlXvBkTfGeDTq79r9_GoMt9&index=27